requirements.txt
tensorflow
pillow
numpy
sv-ttk
customtkinter